Refactor unit action/event classes into files by type
