+++ src/Engine/GameInput/SelectCargo.java
@@ -6,6 +6,7 @@ import java.util.HashMap;
 import Engine.GameAction;
 import Engine.GameActionSet;
 import Engine.XYCoord;
+import Engine.UnitActionLifecycles.UnloadLifecycle;
 import Units.Unit;
 
 /************************************************************
@@ -56,7 +57,7 @@ class SelectCargo extends GameInputState<Object> // Object, not Unit, because "D
     if( option.equals(DONE_OPTION) )
     {
       // Since we don't want to drop any additional units, build the GameAction and move to ActionReady.
-      GameAction ga = new GameAction.UnloadAction(myStateData.unitActor, myStateData.path, myStateData.unitLocationMap);
+      GameAction ga = new UnloadLifecycle.UnloadAction(myStateData.unitActor, myStateData.path, myStateData.unitLocationMap);
 
       // Override the current ActionSet with a new one, since we just redefined it.
       myStateData.actionSet = new GameActionSet( ga, true );
