+++ src/Test/TestVisionMechanics.java
@@ -7,6 +7,9 @@ import Engine.GameAction;
 import Engine.GameScenario;
 import Engine.Path;
 import Engine.Utils;
+import Engine.UnitActionLifecycles.BattleLifecycle;
+import Engine.UnitActionLifecycles.ResupplyLifecycle;
+import Engine.UnitActionLifecycles.WaitLifecycle;
 import Terrain.MapLibrary;
 import Terrain.MapMaster;
 import Terrain.MapWindow;
@@ -71,13 +74,13 @@ public class TestVisionMechanics extends TestCase
     testPassed &= validate(strong.myView.isLocationEmpty(7, 5),   "    We can magically see invisible tanks");
     
     Path foolPath = Utils.findShortestPath(fool, 7, 8, strong.myView);
-    GameAction resupplyBlind = new GameAction.ResupplyAction(fool, foolPath);
+    GameAction resupplyBlind = new ResupplyLifecycle.ResupplyAction(fool, foolPath);
     testPassed &= validate(resupplyBlind.getEvents(testMap).size() == 1, "    Some fool was able to zoom straight through an invisible tank");
 
     Path punchSit = Utils.findShortestPath(punch, punch.x, punch.y, strong.myView);
-    GameAction missBait = new GameAction.AttackAction(strong.myView, punch, punchSit, 6, 5);
+    GameAction missBait = new BattleLifecycle.BattleAction(strong.myView, punch, punchSit, 6, 5);
     testPassed &= validate(missBait.getEvents(testMap).size() == 0, "    You can shoot things hidden in forests.");
-    GameAction missMeat = new GameAction.AttackAction(strong.myView, punch, punchSit, 7, 5);
+    GameAction missMeat = new BattleLifecycle.BattleAction(strong.myView, punch, punchSit, 7, 5);
     testPassed &= validate(missMeat.getEvents(testMap).size() == 0, "    You can shoot invisible things.");
 
     // Drive by the two hidden units
@@ -86,7 +89,7 @@ public class TestVisionMechanics extends TestCase
     excursion.addWaypoint(7, 4);
     excursion.addWaypoint(6, 4);
     excursion.addWaypoint(6, 3);
-    GameAction driveBy = new GameAction.WaitAction(scout, excursion);
+    GameAction driveBy = new WaitLifecycle.WaitAction(scout, excursion);
     testPassed &= validate(driveBy.getEvents(testMap).size() == 1, "    Recons can't move, apparently.");
     performGameAction(driveBy, testMap);
 
@@ -96,12 +99,12 @@ public class TestVisionMechanics extends TestCase
     testPassed &= validate(!strong.myView.isLocationEmpty(6, 5),  "    Doing a drive-by doesn't reveal units in forests");
     testPassed &= validate(!strong.myView.isLocationEmpty(7, 5),  "    Doing a drive-by doesn't reveal invisible tanks");
     
-    GameAction resupplySighted = new GameAction.ResupplyAction(fool, foolPath); // There's no validation for ResupplyAction's constructor, so we'll get pre-empted but generate properly
+    GameAction resupplySighted = new ResupplyLifecycle.ResupplyAction(fool, foolPath); // There's no validation for ResupplyAction's constructor, so we'll get pre-empted but generate properly
     testPassed &= validate(resupplySighted.getEvents(testMap).size() == 1, "    Some fool was able to zoom straight through a visible tank");
     
-    GameAction shootBait = new GameAction.AttackAction(strong.myView, punch, punchSit, 6, 5);
+    GameAction shootBait = new BattleLifecycle.BattleAction(strong.myView, punch, punchSit, 6, 5);
     testPassed &= validate(shootBait.getEvents(testMap).size() == 2, "    You can't shoot things you can see.");
-    GameAction shootMeat = new GameAction.AttackAction(strong.myView, punch, punchSit, 7, 5);
+    GameAction shootMeat = new BattleLifecycle.BattleAction(strong.myView, punch, punchSit, 7, 5);
     testPassed &= validate(shootMeat.getEvents(testMap).size() == 2, "    You can't shoot things you can see.");
     
     // Clean up
