+++ src/Units/Unit.java
@@ -7,7 +7,7 @@ import java.util.Vector;
 import CommandingOfficers.Commander;
 import Engine.GameActionSet;
 import Engine.Path;
-import Engine.UnitActionType;
+import Engine.UnitActionFactory;
 import Engine.XYCoord;
 import Engine.GameEvents.GameEventQueue;
 import Engine.GameEvents.HealUnitEvent;
@@ -251,7 +251,7 @@ public class Unit implements Serializable
   public ArrayList<GameActionSet> getPossibleActions(GameMap map, Path movePath, boolean ignoreResident)
   {
     ArrayList<GameActionSet> actionSet = new ArrayList<GameActionSet>();
-    for( UnitActionType at : model.possibleActions )
+    for( UnitActionFactory at : model.possibleActions )
     {
       GameActionSet actions = at.getPossibleActions(map, movePath, this, ignoreResident);
       if( null != actions )
